Initial setup and added dynamic change on login
